03/05 update

+++ Assets/HealthBar/HealthBar_FillControl.cs
@@ -4,17 +4,17 @@ using UnityEngine.UI;
 public class HealthBar_FillControl : MonoBehaviour
 {
     public Image Bar;
-    public float Fill;
+    public double Fill;
 
     // Start is called before the first frame update
     void Start()
     {
-        Fill = 1f;
+        Fill = 1.0;
     }
 
     // Update is called once per frame
     void Update()
     {
-        Bar.fillAmount = Fill;
+        Bar.fillAmount = (float)Fill;
     }
 }
